refactor: boolean props → union types, fix lint/prettier, add fix scripts
  - SubmitSummary: replace 7 boolean props with token, txStatus, simStatus,
    proposerRole discriminated unions (react-doctor 96→100)
  - AcceptRejectActions: replace 4 boolean props with txStatus union
  - RoleContext: wrap post-hydration setState in startTransition
  - Fix all pre-existing prettier formatting issues across create/ components
  - package.json: fix lint:frontend to include prettier check; add fix,
    fix:frontend, fix:ts, fix:prettier scripts
