+++ package.json
@@ -55,7 +55,11 @@
 		"start:deploy:hardhat": "npm run compile && npm run hardhat:deploy:sepolia",
 		"start:deploy:foundry": "npm run foundry:build && npm run foundry:deploy:sepolia",
 		"lint": "npm run lint:ts && npm run lint:sol && npm run lint:forge && npm run lint:md",
-		"lint:frontend": "cd src && npm run lint:frontend:ts",
+		"lint:frontend": "cd src && npm run lint:frontend",
+		"fix": "npm run fix:ts && npm run fix:prettier && npm run fix:frontend",
+		"fix:frontend": "cd src && npm run fix:frontend",
+		"fix:ts": "eslint --fix 'test/**/*.ts' 'scripts/**/*.ts' 'hardhat.config.ts'",
+		"fix:prettier": "prettier --write .",
 		"lint:sol": "solhint --max-warnings 0 'contracts/src/**/*.sol' 'contracts/test/**/*.sol' 'contracts/script/**/*.sol'",
 		"lint:forge": "cd contracts && forge fmt --check && forge lint",
 		"lint:md": "markdownlint-cli2 \"**/*.md\" \"#contracts/lib\" \"#contracts/out\" \"#contracts/cache\" \"#artifacts\" \"#dist\" \"#node_modules\" \"#src/node_modules\" \"#.next\" \"#src/.next\" \"#.vercel\" \"#.claude\" \"#.serena\" \"#.sixth\" \"#site\" \"#testing\" \"#src/skills\" \"#src/.agents\" \"#src/.claude\"",
