+++ src/components/Input.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { useContext, useId } from "react";
+import { use, useId } from "react";
 
 import { controlClass, FieldIdContext, INPUT_BG } from "@/components/Field";
 
@@ -9,7 +9,7 @@ type InputProps = React.InputHTMLAttributes<HTMLInputElement> & { error?: boolea
 /** Text/number/etc. input that integrates with the surrounding `Field` for label binding and error styling. */
 export function Input({ error = false, className, ...props }: InputProps): React.JSX.Element {
 	const generatedId = useId();
-	const fieldId = useContext(FieldIdContext);
+	const fieldId = use(FieldIdContext);
 	const id = props.id ?? fieldId ?? generatedId;
 	return (
 		<input
