+++ src/doctor.config.ts
@@ -21,6 +21,33 @@ const config = {
 				],
 				rules: ["react-doctor/nextjs-missing-metadata"],
 			},
+			{
+				// Field.tsx intentionally exports its context and utility alongside the
+				// Field component — compound-component pattern. Splitting them out would
+				// add indirection with no meaningful benefit here.
+				files: ["components/Field.tsx"],
+				rules: ["react-doctor/only-export-components"],
+			},
+			{
+				// The fetch here is a fire-and-forget POST (send magic-link email)
+				// after a successful on-chain transaction — not data fetching for
+				// display. Both fetch-in-effect rules are false positives.
+				// Logic moved from page.tsx to useCreatePageState.ts in the hook extraction.
+				files: ["app/create/page.tsx", "app/create/_lib/useCreatePageState.ts"],
+				rules: [
+					"react-doctor/no-fetch-in-effect",
+					"react-doctor/nextjs-no-client-fetch-for-server-data",
+				],
+			},
+			{
+				// RoleContext reads localStorage in a mount effect intentionally.
+				// The useState default must be "freelancer" on both server and client
+				// to avoid React hydration error #418; the effect updates to the
+				// stored value post-hydration. useSyncExternalStore would also work
+				// but is heavier for this simple use case.
+				files: ["contexts/RoleContext.tsx"],
+				rules: ["react-doctor/no-initialize-state"],
+			},
 		],
 	},
 };
